RR-939 - Move aria text resolver so it can be shared by non-induction related controllers
