+++ server/routes/dynamicAriaTextResolver.ts
