+++ server/routes/induction/create/previousWorkExperienceDetailCreateController.ts
@@ -3,7 +3,7 @@ import { NextFunction, Request, RequestHandler, Response } from 'express'
 import type { InductionDto } from 'inductionDto'
 import PreviousWorkExperienceDetailController from '../common/previousWorkExperienceDetailController'
 import { getPreviousPage } from '../../pageFlowHistory'
-import getDynamicBackLinkAriaText from '../dynamicAriaTextResolver'
+import getDynamicBackLinkAriaText from '../../dynamicAriaTextResolver'
 import TypeOfWorkExperienceValue from '../../../enums/typeOfWorkExperienceValue'
 import validatePreviousWorkExperienceDetailForm from '../../validators/induction/previousWorkExperienceDetailFormValidator'
 import { getNextPage, isLastPage } from '../../pageFlowQueue'
