+++ server/routes/induction/create/previousWorkExperienceTypesCreateController.ts
@@ -4,7 +4,7 @@ import type { InductionDto } from 'inductionDto'
 import type { PreviousWorkExperienceTypesForm } from 'inductionForms'
 import PreviousWorkExperienceTypesController from '../common/previousWorkExperienceTypesController'
 import { buildNewPageFlowHistory, getPreviousPage } from '../../pageFlowHistory'
-import getDynamicBackLinkAriaText from '../dynamicAriaTextResolver'
+import getDynamicBackLinkAriaText from '../../dynamicAriaTextResolver'
 import validatePreviousWorkExperienceTypesForm from '../../validators/induction/previousWorkExperienceTypesFormValidator'
 import previousWorkExperienceTypeScreenOrderComparator from '../previousWorkExperienceTypeScreenOrderComparator'
 import logger from '../../../../logger'
