+++ server/routes/induction/create/wantToAddQualificationsCreateController.ts
@@ -2,7 +2,7 @@ import { Request, RequestHandler, Response } from 'express'
 import type { InductionDto } from 'inductionDto'
 import WantToAddQualificationsController from '../common/wantToAddQualificationsController'
 import YesNoValue from '../../../enums/yesNoValue'
-import getDynamicBackLinkAriaText from '../dynamicAriaTextResolver'
+import getDynamicBackLinkAriaText from '../../dynamicAriaTextResolver'
 import validateWantToAddQualificationsForm from '../../validators/induction/wantToAddQualificationsFormValidator'
 import EducationLevelValue from '../../../enums/educationLevelValue'
 import { getPreviousPage } from '../../pageFlowHistory'
