+++ server/routes/induction/create/workedBeforeCreateController.ts
@@ -2,7 +2,7 @@ import { NextFunction, Request, RequestHandler, Response } from 'express'
 import type { WorkedBeforeForm } from 'inductionForms'
 import type { InductionDto } from 'inductionDto'
 import WorkedBeforeController from '../common/workedBeforeController'
-import getDynamicBackLinkAriaText from '../dynamicAriaTextResolver'
+import getDynamicBackLinkAriaText from '../../dynamicAriaTextResolver'
 import validateWorkedBeforeForm from '../../validators/induction/workedBeforeFormValidator'
 import { getPreviousPage } from '../../pageFlowHistory'
 import HasWorkedBeforeValue from '../../../enums/hasWorkedBeforeValue'
