+++ server/routes/induction/update/highestLevelOfEducationUpdateController.ts
@@ -6,7 +6,7 @@ import validateHighestLevelOfEducationForm from '../../validators/induction/high
 import toCreateOrUpdateInductionDto from '../../../data/mappers/createOrUpdateInductionDtoMapper'
 import logger from '../../../../logger'
 import { getPreviousPage } from '../../pageFlowHistory'
-import getDynamicBackLinkAriaText from '../dynamicAriaTextResolver'
+import getDynamicBackLinkAriaText from '../../dynamicAriaTextResolver'
 
 /**
  * Controller for the Update of the Highest Level of Education screen of the Induction.
