+++ server/routes/induction/update/previousWorkExperienceDetailUpdateController.ts
@@ -9,7 +9,7 @@ import validatePreviousWorkExperienceDetailForm from '../../validators/induction
 import TypeOfWorkExperienceValue from '../../../enums/typeOfWorkExperienceValue'
 import { getNextPage, isLastPage } from '../../pageFlowQueue'
 import { getPreviousPage } from '../../pageFlowHistory'
-import getDynamicBackLinkAriaText from '../dynamicAriaTextResolver'
+import getDynamicBackLinkAriaText from '../../dynamicAriaTextResolver'
 
 /**
  * Controller for the Update of the Previous Work Experience Detail screen of the Induction.
